pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "insighta"
version = "0.1.0"
dependencies = [
    "click",
    "httpx",
    "rich",
    "python-dotenv"
]

[project.scripts]
insighta = "insighta.cli:app"